refactor llm to package
